refactor(chromium,firefox): add some api feat(timestamp): support chromium and firefox

+++ crates/decrypt-cookies/src/chromium/items/cookie/entities/cookies.rs
@@ -8,7 +8,7 @@ use sea_orm::entity::prelude::*;
 #[derive(PartialEq, Eq)]
 #[sea_orm(table_name = "cookies")]
 pub struct Model {
-    pub creation_utc:       i32,
+    pub creation_utc:       i64,
     pub host_key:           String,
     pub top_frame_site_key: String,
     pub name:               String,
@@ -17,17 +17,17 @@ pub struct Model {
     #[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
     pub encrypted_value:    Vec<u8>,
     pub path:               String,
-    pub expires_utc:        i32,
+    pub expires_utc:        i64,
     pub is_secure:          i32,
     pub is_httponly:        i32,
-    pub last_access_utc:    i32,
+    pub last_access_utc:    i64,
     pub has_expires:        i32,
     pub is_persistent:      i32,
     pub priority:           i32,
     pub samesite:           i32,
     pub source_scheme:      i32,
     pub source_port:        i32,
-    pub last_update_utc:    i32,
+    pub last_update_utc:    i64,
     // pub is_same_party:      i32,
 }
